requirements.txt
Flask==0.12
Flask-Cors==3.0.4
Flask-Login==0.4.1
SQLAlchemy==1.2.7
Werkzeug==0.11.15
asn1crypto==0.24.0
cryptography==2.2.2
pycrypto==2.6.1
pycryptodome==3.8.1
